requirements.txt
scikit-learn>=1.0
joblib
flask
numpy
pillow